Top nav bar and breadcrumbs!!
I added a method run before every handler that builds "breadcrumbs" for the page that is being viewed. If any are generated by the method, they will be displayed in the nav bar if they are formatted correctly. View buildBreadCrumbs() in app.js and /games/ and /games/id for examples