Breaking super_spectator, adding 960
So I'm in the middle of some stuff, but I wanted to add a couple fixes
and a couple other things. So I added choosing a game mode. The only
ones implemented are classic and 960, the other two don't do anything
yet. I broke super_spectator mode because I'm changing how it works, so
we'll see how that goes